requirements.txt
# Data Processing
pandas==2.0.3
numpy==1.24.3
pyarrow==12.0.1
openpyxl==3.1.2

# PySpark
pyspark==3.4.1
py4j==0.10.9.7

# Database
psycopg2-binary==2.9.7
sqlalchemy==1.4.49
pymongo==4.4.1

# Airflow
apache-airflow==2.7.1
apache-airflow-providers-postgres==5.6.0
apache-airflow-providers-amazon==8.5.1

# SQL & Templates
jinja2==3.1.2
jinjasql==0.1.8

# Testing
pytest==7.4.2
faker==19.6.1
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
rich==13.5.2
tabulate==0.9.0

# Notebook
jupyter==1.0.0
ipykernel==6.25.1